Continued with base functions, some of them not finished. CheckCollision is only a skeleton. Made many variables final, replaced maxSpeed to Cell, and communication package under network. Finished mapObjects so it can be sued for communication.

+++ AgarIoClone/src/server/model/helper/Coords.java
@@ -10,11 +10,11 @@ public class Coords {
         this.y=y;
     }
     
-    public void setX(int x){
+    public void setX(float x){
         this.x=x;
     }
     
-    public void setY(int y){
+    public void setY(float y){
         this.y=y;
     }
